fixed signout button, was working on main by acc
