+++ app/_layout.tsx
@@ -53,6 +53,7 @@ function RootLayoutNav() {
         <Stack.Screen name="index" options={{ headerShown: false }} />
         <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
         <Stack.Screen name="modal" options={{ presentation: 'modal' }} />
+        <Stack.Screen name="resetToRoot" />
       </Stack>
     </ThemeProvider>
   );
