+++ app/resetToRoot.tsx
@@ -1,0 +1,15 @@
+import { useRouter } from "expo-router";
+    import { useEffect } from "react";
+    
+    const ResetToRoot = () => {
+      const router = useRouter();
+    
+      useEffect(() => {
+        setTimeout(() => {
+          router.replace("/");
+        }, 0);
+      });
+      return <></>;
+    };
+    
+    export default ResetToRoot;
