+++ package.json
@@ -18,8 +18,10 @@
     "@react-native-async-storage/async-storage": "^1.24.0",
     "@react-navigation/native": "^7.1.6",
     "axios": "^1.9.0",
+    "dotenv": "^16.5.0",
     "expo": "~53.0.9",
     "expo-build-properties": "~0.14.6",
+    "expo-constants": "~17.1.6",
     "expo-font": "~13.3.1",
     "expo-linking": "~7.1.5",
     "expo-location": "~18.1.5",
@@ -32,6 +34,7 @@
     "react": "19.0.0",
     "react-dom": "19.0.0",
     "react-native": "0.79.2",
+    "react-native-dotenv": "^3.4.11",
     "react-native-google-places-autocomplete": "^2.5.7",
     "react-native-maps": "1.20.1",
     "react-native-reanimated": "~3.17.4",
@@ -42,6 +45,7 @@
   "devDependencies": {
     "@babel/core": "^7.25.2",
     "@types/react": "~19.0.10",
+    "@types/react-native": "^0.72.8",
     "jest": "^29.2.1",
     "jest-expo": "~53.0.5",
     "react-test-renderer": "19.0.0",
